Add 'networking.k8s.io apiGroup in ingress roles
Current fury-kubernetes-ingress release is missing networking.k8s.io
apiGroup in ingress-related resource in RBAC role. As a result this is
preventing the pods from starting up. This commit is adding the missing
apiGroup.
Together with this I've re-organized the list in order to put the
ingress-related resources close to each-other - as it's in the ingress
upstream:
https://github.com/kubernetes/ingress-nginx/blob/master/deploy/static/mandatory.yaml